backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Providers
openai==1.3.7
google-generativeai==0.3.1

# Vector Database & Embeddings
qdrant-client==1.7.0
sentence-transformers==2.5.0
chromadb==0.4.18

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & WebSocket
httpx==0.25.2
websockets==12.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
pyyaml==6.0.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Data Processing
numpy==1.26.2
pandas==2.1.4
tiktoken==0.5.2
